+++ apps/ymir-sdl3/src/app/ui/defs/settings_defs.hpp
@@ -2,6 +2,6 @@
 
 namespace app::ui {
 
-enum class SettingsTab { None, General, Hotkeys, System, IPL, Input, Video, Audio, Cartridge, CDBlock };
+enum class SettingsTab { None, General, Hotkeys, System, IPL, Input, Video, Audio, Cartridge, CDBlock, Tweaks };
 
 } // namespace app::ui