+++ apps/ymir-sdl3/src/app/ui/views/settings/audio_settings_view.cpp
@@ -3,6 +3,9 @@
 #include <app/events/emu_event_factory.hpp>
 
 #include <app/ui/widgets/common_widgets.hpp>
+#include <app/ui/widgets/settings_widgets.hpp>
+
+#include <RtMidi.h>
 
 using namespace ymir;
 
@@ -16,8 +19,6 @@ AudioSettingsView::AudioSettingsView(SharedContext &context)
 void AudioSettingsView::Display() {
     auto &settings = m_context.settings.audio;
     auto &config = m_context.saturn.configuration.audio;
-
-    auto *drawList = ImGui::GetWindowDrawList();
 
     // -----------------------------------------------------------------------------------------------------------------
 
@@ -39,27 +40,11 @@ void AudioSettingsView::Display() {
 
     // -----------------------------------------------------------------------------------------------------------------
 
-    using InterpMode = core::config::audio::SampleInterpolationMode;
-
-    auto interpOption = [&](const char *name, InterpMode mode) {
-        const std::string label = fmt::format("{}##sample_interp", name);
-        ImGui::SameLine();
-        if (MakeDirty(ImGui::RadioButton(label.c_str(), config.interpolation == mode))) {
-            config.interpolation = mode;
-        }
-    };
-
     ImGui::PushFont(m_context.fonts.sansSerif.large.bold);
     ImGui::SeparatorText("Quality");
     ImGui::PopFont();
 
-    ImGui::AlignTextToFramePadding();
-    ImGui::TextUnformatted("Interpolation:");
-    widgets::ExplanationTooltip("- Nearest neighbor: Cheapest option with grittier sounds.\n"
-                                "- Linear: Hardware accurate option with softer sounds. (default)",
-                                m_context.displayScale);
-    interpOption("Nearest neighbor", InterpMode::NearestNeighbor);
-    interpOption("Linear", InterpMode::Linear);
+    widgets::settings::audio::InterpolationMode(m_context);
 
     // -----------------------------------------------------------------------------------------------------------------
 
@@ -152,63 +137,7 @@ void AudioSettingsView::Display() {
     ImGui::SeparatorText("Accuracy");
     ImGui::PopFont();
 
-    int stepGranularity = settings.stepGranularity;
-    if (MakeDirty(ImGui::SliderInt("Emulation step granularity", &stepGranularity, 0, 5, "%d",
-                                   ImGuiSliderFlags_AlwaysClamp))) {
-        // m_context.EnqueueEvent(events::emu::SetSCSPStepGranularity(stepGranularity));
-        settings.stepGranularity = stepGranularity;
-    }
-    widgets::ExplanationTooltip(
-        "Adjusts emulation granularity for the SCSP.\n"
-        "\n"
-        "Increasing this setting causes the SCSP to be emulated in smaller timeslices (up to 32 times as often as "
-        "sample-level processing), significantly lowering performance in exchange for a higher level of accuracy that "
-        "doesn't benefit the vast majority of commercial games.\n"
-        "\n"
-        "This option might be of interest to homebrew developers who need extra accuracy in some way.",
-        m_context.displayScale);
-
-    {
-        static constexpr ImU32 kGraphBorderColor = 0xE0BBE0F0;
-        static constexpr ImU32 kGraphBackgroundColor = 0xAA253840;
-        static constexpr ImU32 kGraphSliceFillColor = 0xE04AC3F7;
-        static constexpr ImU32 kGraphSliceFillColorAlt = 0xE02193C4;
-        static constexpr ImU32 kGraphSlotSeparatorColor = 0xE02A6F8C;
-
-        const auto basePos = ImGui::GetCursorScreenPos();
-        const auto avail = ImGui::GetContentRegionAvail();
-        const auto innerSpacing = ImGui::GetStyle().ItemInnerSpacing;
-        const auto framePadding = ImGui::GetStyle().FramePadding;
-        const float textHeight = ImGui::GetTextLineHeightWithSpacing();
-        const float graphWidth = ImGui::CalcItemWidth();
-        const float graphHeight = ImGui::GetFrameHeight();
-        const float borderThickness = 2.0f * m_context.displayScale;
-        const float sliceWidth = graphWidth / (1 << stepGranularity);
-        const float slotWidth = graphWidth / 32.0f;
-        const float sepThickness = 1.5f * m_context.displayScale;
-
-        ImGui::Dummy(ImVec2(graphWidth, graphHeight));
-        drawList->AddRectFilled(basePos, ImVec2(basePos.x + graphWidth, basePos.y + graphHeight),
-                                kGraphBackgroundColor);
-        for (uint32 i = 0; i < (1 << stepGranularity); ++i) {
-            const float xStart = basePos.x + i * sliceWidth;
-            const float xEnd = xStart + sliceWidth;
-            drawList->AddRectFilled(ImVec2(xStart, basePos.y), ImVec2(xEnd, basePos.y + graphHeight),
-                                    (i & 1) ? kGraphSliceFillColorAlt : kGraphSliceFillColor);
-        }
-        for (uint32 i = 1; i < 32; ++i) {
-            const float x = basePos.x + i * slotWidth;
-            drawList->AddLine(ImVec2(x, basePos.y), ImVec2(x, basePos.y + graphHeight), kGraphSlotSeparatorColor,
-                              sepThickness);
-        }
-        drawList->AddRect(basePos, ImVec2(basePos.x + graphWidth, basePos.y + graphHeight), kGraphBorderColor, 0.05, 0,
-                          borderThickness);
-        drawList->AddText(ImVec2(basePos.x + graphWidth + innerSpacing.x, basePos.y + framePadding.y), 0xFFFFFFFF,
-                          fmt::format("Step size: {} {}{}", (32 >> stepGranularity),
-                                      (stepGranularity < 5 ? "slots" : "slot"),
-                                      (stepGranularity == 0 ? " (one sample)" : ""))
-                              .c_str());
-    }
+    widgets::settings::audio::StepGranularity(m_context);
 
     // -----------------------------------------------------------------------------------------------------------------