+++ apps/ymir-sdl3/src/app/ui/views/settings/audio_settings_view.hpp
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <RtMidi.h>
-
 #include "settings_view_base.hpp"
 
 namespace app::ui {