+++ apps/ymir-sdl3/src/app/ui/views/settings/tweaks_settings_view.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+#include "settings_view_base.hpp"
+
+namespace app::ui {
+
+class TweaksSettingsView : public SettingsViewBase {
+public:
+    TweaksSettingsView(SharedContext &context);
+
+    void Display();
+
+private:
+    void DisplayEnhancements();
+    void DisplayAccuracyOptions();
+};
+
+} // namespace app::ui