+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - App: Added option to create internal backup RAM files per game. (#99)
 - App: Added option to override UI scale. (#251)
 - App: Added option to toggle fullscreen by double-clicking the display. (#197)
+- App: Added new Tweaks tab to Settings window consolidating all accuracy, compatibility and enhancement settings.
 - App: Added recent games list to File menu. (#196)
 - App: Automatically center Settings window when opening it. (#251)
 - App: Close windows when pressing B or Circle on gamepads while nothing is focused. (#251)